fix: 드롭다운 필터를 훅으로 분리하여 hook 간 순서 보장이 되지 않는 문제 해결 (롤백)
